feature：创建订单补充库存校验和订单超时处理

+++ cloudfarming-service/biz-order/src/main/java/com/vv/cloudfarming/order/mq/modal/MultiDelayMessage.java
@@ -19,15 +19,15 @@ public class MultiDelayMessage<T> {
     /**
      * 记录延迟时间的集合
      */
-    private List<Long> delayMillis;
+    private List<Integer> delayMillis;
 
     public MultiDelayMessage() {}
 
-    public MultiDelayMessage(T data, List<Long> delayMillis) {
+    public MultiDelayMessage(T data, List<Integer> delayMillis) {
         this.data = data;
         this.delayMillis = delayMillis;
     }
-    public static <T> MultiDelayMessage<T> of(T data, Long ... delayMillis){
+    public static <T> MultiDelayMessage<T> of(T data, Integer ... delayMillis){
         return new MultiDelayMessage<>(data, CollUtil.newArrayList(delayMillis));
     }
 
@@ -35,7 +35,7 @@ public class MultiDelayMessage<T> {
      * 获取并移除下一个延迟时间
      * @return 队列中的第一个延迟时间
      */
-    public Long removeNextDelay(){
+    public Integer removeNextDelay(){
         return delayMillis.remove(0);
     }
